+++ packages/uploadthing/src/client.ts
@@ -1,14 +1,14 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { pollForFileData, UploadThingError } from "@uploadthing/shared";
+import { safeParseJSON, UploadThingError } from "@uploadthing/shared";
 
-import { maybeParseResponseXML } from "./internal/s3-error-parser";
+import type { UploadThingResponse } from "./internal/handler";
+import { uploadPartWithProgress } from "./internal/multi-part";
 import type {
-  ActionType,
+  DistributiveOmit,
   FileRouter,
   inferEndpointInput,
+  inferEndpointOutput,
 } from "./internal/types";
+import { createAPIRequestUrl, createUTReporter } from "./internal/ut-reporter";
 
 /**
  * @internal
@@ -16,114 +16,74 @@ import type {
  */
 export * from "./internal/component-theming";
 
-function fetchWithProgress(
-  url: string,
-  opts: {
-    headers?: Headers;
-    method?: string;
-    body?: string | FormData;
-  } = {},
-  onProgress?: (this: XMLHttpRequest, progress: ProgressEvent) => void,
-  onUploadBegin?: (this: XMLHttpRequest, progress: ProgressEvent) => void,
-) {
-  return new Promise<XMLHttpRequest>((res, rej) => {
-    const xhr = new XMLHttpRequest();
-    xhr.open(opts.method ?? "get", url);
-    opts.headers &&
-      Object.keys(opts.headers).forEach(
-        (h) =>
-          opts.headers && xhr.setRequestHeader(h, opts.headers.get(h) ?? ""),
-      );
-    xhr.onload = (e) => {
-      res(e.target as XMLHttpRequest);
-    };
+type UploadFilesOptions<
+  TRouter extends FileRouter,
+  TEndpoint extends keyof TRouter,
+> = {
+  onUploadProgress?: ({
+    file,
+    progress,
+  }: {
+    file: string;
+    progress: number;
+  }) => void;
+  onUploadBegin?: ({ file }: { file: string }) => void;
 
-    xhr.onerror = rej;
-    if (xhr.upload && onProgress) xhr.upload.onprogress = onProgress;
-    if (xhr.upload && onUploadBegin) xhr.upload.onloadstart = onUploadBegin;
-    xhr.send(opts.body);
+  files: File[];
+
+  /**
+   * URL to the UploadThing API endpoint
+   * @example URL { http://localhost:3000/api/uploadthing }
+   * @example URL { https://www.example.com/api/uploadthing }
+   */
+  url: URL;
+} & (undefined extends inferEndpointInput<TRouter[TEndpoint]>
+  ? // eslint-disable-next-line @typescript-eslint/ban-types
+    {}
+  : {
+      input: inferEndpointInput<TRouter[TEndpoint]>;
+    });
+
+export const INTERNAL_DO_NOT_USE__fatalClientError = (e: Error) =>
+  new UploadThingError({
+    code: "INTERNAL_CLIENT_ERROR",
+    message: "Something went wrong. Please report this to UploadThing.",
+    cause: e,
   });
-}
 
-const createAPIRequestUrl = (config: {
-  url?: string;
-  slug: string;
-  actionType: ActionType;
-}) => {
-  const url = new URL(
-    config.url ?? `${window.location.origin}/api/uploadthing`,
-  );
-
-  const queryParams = new URLSearchParams(url.search);
-  queryParams.set("actionType", config.actionType);
-  queryParams.set("slug", config.slug);
-
-  url.search = queryParams.toString();
-  return url.toString();
-};
-
-type UploadFilesOptions<TRouter extends FileRouter> = {
-  [TEndpoint in keyof TRouter]: {
-    endpoint: TEndpoint;
-    onUploadProgress?: ({
-      file,
-      progress,
-    }: {
-      file: string;
-      progress: number;
-    }) => void;
-    onUploadBegin?: ({ file }: { file: string }) => void;
-    input?: inferEndpointInput<TRouter[TEndpoint]>;
-
-    files: File[];
-  };
-}[keyof TRouter];
-
-export type UploadFileResponse = {
-  /**
-   * @deprecated
-   * use `name` instead
-   */
-  fileName: string;
+export type UploadFileResponse<TServerOutput> = {
   name: string;
-  /**
-   * @deprecated
-   * use `size` instead
-   */
-  fileSize: number;
   size: number;
-  /**
-   * @deprecated
-   * use `key` instead
-   */
-  fileKey: string;
   key: string;
-  /**
-   * @deprecated
-   * use `url` instead
-   */
-  fileUrl: string;
   url: string;
+  // Matches what's returned from the serverside `onUploadComplete` callback
+  serverData: TServerOutput;
 };
 
-export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
-  opts: UploadFilesOptions<TRouter>,
-  config?: {
-    url?: string;
-  },
+export const DANGEROUS__uploadFiles = async <
+  TRouter extends FileRouter,
+  TEndpoint extends keyof TRouter,
+>(
+  endpoint: TEndpoint,
+  opts: UploadFilesOptions<TRouter, TEndpoint>,
 ) => {
+  const reportEventToUT = createUTReporter({
+    endpoint: String(endpoint),
+    url: opts.url,
+  });
+
   // Get presigned URL for S3 upload
   const s3ConnectionRes = await fetch(
     createAPIRequestUrl({
-      url: config?.url,
-      slug: String(opts.endpoint),
+      url: opts.url,
+      slug: String(endpoint),
       actionType: "upload",
     }),
     {
       method: "POST",
       body: JSON.stringify({
-        files: opts.files.map((f) => f.name),
-        input: opts.input,
+        input: "input" in opts ? opts.input : null,
+        files: opts.files.map((f) => ({ name: f.name, size: f.size })),
       }),
       // Express requires Content-Type to be explicitly set to parse body properly
       headers: {
@@ -137,18 +97,15 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
       throw error;
     }
 
-    // attempt to parse response
-    try {
-      return res.json();
-    } catch (e) {
-      // response is not JSON
-      console.error(e);
+    const jsonOrError = await safeParseJSON<UploadThingResponse>(res);
+    if (jsonOrError instanceof Error) {
       throw new UploadThingError({
         code: "BAD_REQUEST",
-        message: `Failed to parse response as JSON. Got: ${await res.text()}`,
-        cause: e,
+        message: jsonOrError.message,
+        cause: res,
       });
     }
+    return jsonOrError;
   });
 
   if (!s3ConnectionRes || !Array.isArray(s3ConnectionRes)) {
@@ -159,8 +116,8 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
     });
   }
 
-  const fileUploadPromises = s3ConnectionRes.map(async (presigned: any) => {
-    const file = opts.files.find((f) => f.name === presigned.name);
+  const fileUploadPromises = s3ConnectionRes.map(async (presigned) => {
+    const file = opts.files.find((f) => f.name === presigned.fileName);
 
     if (!file) {
       console.error("No file found for presigned URL", presigned);
@@ -168,114 +125,111 @@ export const DANGEROUS__uploadFiles = async <TRouter extends FileRouter>(
         code: "NOT_FOUND",
         message: "No file found for presigned URL",
         cause: `Expected file with name ${
-          presigned.name
+          presigned.fileName
         } but got '${opts.files.join(",")}'`,
       });
     }
-    const { url, fields } = presigned.presignedUrl;
-    const formData = new FormData();
 
-    // Give content type to blobs because S3 is dumb
-    // check if content-type is one of the allowed types, or if not and blobs are allowed, use application/octet-stream
-    if (
-      presigned.fileType === file.type.split("/")[0] ||
-      presigned.fileType === file.type
-    ) {
-      formData.append("Content-Type", file.type);
-    } else if (presigned.fileType === "blob") {
-      formData.append("Content-Type", "application/octet-stream");
-    } else if (presigned.fileType === "pdf") {
-      formData.append("Content-Type", "application/pdf");
+    const { presignedUrls, uploadId, chunkSize, contentDisposition, key } =
+      presigned;
+
+    let uploadedBytes = 0;
+
+    let etags: { tag: string; partNumber: number }[];
+    try {
+      etags = await Promise.all(
+        presignedUrls.map(async (url, index) => {
+          const offset = chunkSize * index;
+          const end = Math.min(offset + chunkSize, file.size);
+          const chunk = file.slice(offset, end);
+
+          const etag = await uploadPartWithProgress({
+            url,
+            chunk: chunk,
+            contentDisposition,
+            fileType: file.type,
+            fileName: file.name,
+            maxRetries: 10,
+            onProgress: (delta) => {
+              uploadedBytes += delta;
+              const percent = (uploadedBytes / file.size) * 100;
+              opts.onUploadProgress?.({ file: file.name, progress: percent });
+            },
+          });
+
+          return { tag: etag, partNumber: index + 1 };
+        }),
+      );
+    } catch (error) {
+      await reportEventToUT("failure", {
+        fileKey: key,
+        uploadId,
+        fileName: file.name,
+        s3Error: (error as Error).toString(),
+      });
+      throw "unreachable"; // failure event will throw for us
     }
 
-    // Dump all values from response (+ the file itself) into form for S3 upload
-    Object.entries({ ...fields, file: file }).forEach(([key, value]) => {
-      formData.append(key, value as Blob);
+    // Tell the server that the upload is complete
+    const uploadOk = await reportEventToUT("multipart-complete", {
+      uploadId,
+      fileKey: key,
+      etags,
     });
+    if (!uploadOk) {
+      console.log("Failed to alert UT of upload completion");
+      throw new UploadThingError({
+        code: "UPLOAD_FAILED",
+        message: "Failed to alert UT of upload completion",
+      });
+    }
 
-    // Do S3 upload
-    const upload = await fetchWithProgress(
-      url,
-      {
-        method: "POST",
-        body: formData,
-        headers: new Headers({
-          Accept: "application/xml",
-        }),
-      },
-      (progressEvent) =>
-        opts.onUploadProgress?.({
-          file: file.name,
-          progress: (progressEvent.loaded / progressEvent.total) * 100,
-        }),
-      () => {
-        opts.onUploadBegin?.({
-          file: file.name,
-        });
-      },
+    const serverData = await fetch(opts.url, {
+      headers: { "x-uploadthing-polling-key": key },
+    }).then(
+      (res) => res.json() as Promise<inferEndpointOutput<TRouter[TEndpoint]>>,
     );
 
-    if (upload.status > 299 || upload.status < 200) {
-      // tell uploadthing infra server that upload failed
-      await fetch(
-        createAPIRequestUrl({
-          url: config?.url,
-          slug: String(opts.endpoint),
-          actionType: "failure",
-        }),
-        {
-          method: "POST",
-          body: JSON.stringify({
-            fileKey: fields.key,
-          }),
-        },
-      );
-
-      // Attempt to parse response as XML
-      const parsed = maybeParseResponseXML(upload.responseText);
-
-      // Throw an error for the client
-      if (parsed?.message) {
-        throw new UploadThingError({
-          code: parsed.code,
-          message: parsed.message,
-        });
-      } else {
-        throw new UploadThingError({
-          code: "UPLOAD_FAILED",
-          message: `Failed to upload file ${file.name} to S3`,
-          cause: upload.responseText,
-        });
-      }
-    }
-
-    // Generate a URL for the uploaded image since AWS won't give me one
-    const genUrl = "https://utfs.io/f/" + encodeURIComponent(fields.key);
-
-    // Poll for file data, this way we know that the client-side onUploadComplete callback will be called after the server-side version
-    await pollForFileData(presigned.key);
-
-    // TODO: remove `file` prefix in next major version
-    const ret: UploadFileResponse = {
-      fileName: file.name,
+    return {
       name: file.name,
-      fileSize: file.size,
       size: file.size,
-      fileKey: presigned.key,
       key: presigned.key,
-      fileUrl: genUrl,
-      url: genUrl,
-    };
-    return ret;
+
+      serverData,
+      url: "https://utfs.io/f/" + key,
+    } satisfies UploadFileResponse<inferEndpointOutput<TRouter[TEndpoint]>>;
   });
 
   return Promise.all(fileUploadPromises);
 };
 
-export const genUploader = <
-  TRouter extends FileRouter,
->(): typeof DANGEROUS__uploadFiles<TRouter> => {
-  return DANGEROUS__uploadFiles;
+export const genUploader = <TRouter extends FileRouter>(initOpts: {
+  /**
+   * URL to the UploadThing API endpoint
+   * @example URL { /api/uploadthing }
+   * @example URL { https://www.example.com/api/uploadthing }
+   *
+   * If relative, host will be inferred from either the `VERCEL_URL` environment variable or `window.location.origin`
+   *
+   * @default (VERCEL_URL ?? window.location.origin) + "/api/uploadthing"
+   */
+  url?: string | URL;
+}) => {
+  const url =
+    initOpts?.url instanceof URL ? initOpts.url : getFullApiUrl(initOpts?.url);
+
+  return <TEndpoint extends keyof TRouter>(
+    endpoint: TEndpoint,
+    opts: DistributiveOmit<
+      Parameters<typeof DANGEROUS__uploadFiles<TRouter, TEndpoint>>[1],
+      "url"
+    >,
+  ) =>
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+    DANGEROUS__uploadFiles<TRouter, TEndpoint>(endpoint, {
+      ...opts,
+      url,
+    } as any);
 };
 
 export const classNames = (...classes: (string | boolean)[]) => {
@@ -303,3 +257,39 @@ export const generateClientDropzoneAccept = (fileTypes: string[]) => {
 
   return Object.fromEntries(mimeTypes.map((type) => [type, []]));
 };
+
+// Returns a full URL to the dev's uploadthing endpoint
+export function getFullApiUrl(maybeUrl?: string): URL {
+  const base = (() => {
+    if (typeof window !== "undefined") {
+      return window.location.origin;
+    }
+
+    // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
+    if (typeof process !== "undefined" && process?.env?.VERCEL_URL) {
+      return `https://${process.env.VERCEL_URL}`;
+    }
+
+    // @ts-expect-error - import meta is not defined in node
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    if (import.meta.env?.VERCEL_URL) {
+      // @ts-expect-error - import meta is not defined in node
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      return `https://${import.meta.env.VERCEL_URL}`;
+    }
+
+    return "http://localhost:3000";
+  })();
+
+  try {
+    const url = new URL(maybeUrl ?? "/api/uploadthing", base);
+    if (url.pathname === "/") {
+      url.pathname = "/api/uploadthing";
+    }
+    return url;
+  } catch (err) {
+    throw new Error(
+      `Failed to parse '${maybeUrl}' as a URL. Make sure it's a valid URL or path`,
+    );
+  }
+}