+++ packages/uploadthing/src/constants.ts
@@ -1,6 +1,4 @@
-import type { PackageJson } from "type-fest";
+import packageJson from "../package.json";
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const packageJson = require("../package.json") as PackageJson;
 if (!packageJson.version) throw new Error("no version found in package.json");
 export const UPLOADTHING_VERSION = packageJson.version;