+++ packages/uploadthing/src/h3.ts
@@ -1,3 +1,4 @@
+import EventEmitter from "events";
 import type { H3Event } from "h3";
 import {
   assertMethod,
@@ -36,7 +37,8 @@ export const createUploadthing = <TErrorShape extends Json>(
 export const createH3EventHandler = <TRouter extends FileRouter>(
   opts: RouterWithConfig<TRouter>,
 ) => {
-  const requestHandler = buildRequestHandler(opts);
+  const ee = new EventEmitter();
+  const requestHandler = buildRequestHandler(opts, ee);
   const getBuildPerms = buildPermissionsInfoHandler<TRouter>(opts);
 
   return defineEventHandler(async (event) => {
@@ -45,6 +47,16 @@ export const createH3EventHandler = <TRouter extends FileRouter>(
 
     // GET
     if (event.method === "GET") {
+      const clientPollingKey =
+        getRequestHeaders(event)["x-uploadthing-polling"];
+      if (clientPollingKey) {
+        const eventData = await new Promise((resolve) => {
+          ee.addListener("callbackDone", resolve);
+        });
+        ee.removeAllListeners("callbackDone");
+        return eventData;
+      }
+
       return getBuildPerms();
     }